Added the "Rope Cutting" question.
In this quesition you need to get the parameter "Rope Length" and cut the length in parts. The parts have to be the biggest value that they can on multiplication.
You SHOULDN'T bruteforce this question but cutting the length in every way possible since that would take a much longer time than desired.
The answer is simple you just have to look at the smallest multiplicator of every number.

Hint: Every number can be made of 2s and 3s.
  \_ Example:  7->3+2+2     8->3+3+2     9->3+3+3     10->3+3+2+2

Every number can also be reached by using the number 1 but the question wants to get the BIGGEST value from MULTIPLICATION.
So lets take the number 8 for example.
  \_ 8->3+3+2       &  8->1+1+1+1+1+1+1+1
  |_        3*3*2=18  &  1*1*1*1*1*1*1*1=1

So creating the same number from ones will ALWAYS give us the smallest value. (Not implementing things like: 3->1+1+1+0)