New question added "Subarray With Given Sum".
This one is pretty straightforward. One of the parameters takes an array the other one the wanted sum.
You're supposed to code it so it finds the subarray which has a sum that equals to the wanted sum.

Basic example:
X = Start Index
Y = End Index
X has to be smaller than Y since the start index can't be bigger than the end index.
You slice the array testing EVERY end index with every start index until you find the match.
If you dont find a match you return None.