Added new question "Count The Triplets".
This is the most simple so far since its not that specific. We just check A + B is in the array and if it is we just add one to our count.
The first loop's value must be bigger than the second loop's value because: 
array[X] + array[Y] = array[Y] + array[X]
If we don't put that restraint then we will always get twice the amount. (I believe that you can just do "return count / 2" but I haven't tested it.)